% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{accrossAssessorReport}
\alias{accrossAssessorReport}
\title{accrossAssessorReport}
\usage{
accrossAssessorReport(semAllAssessordf, nAssesor, nParticipant)
}
\arguments{
\item{semAllAssessordf}{\link{dataframe} standard error of measurement table computed for ALL assessors}

\item{nAssesor}{\link{integer} number of assessors}

\item{nParticipant}{\link{integer} number of participants}
}
\value{
[]
}
\description{
Report of across asessor differences
}
\section{Warning}{

}

