% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMetrics.R
\name{computeMetrics}
\alias{computeMetrics}
\title{compute metrics from the Metrics package}
\usage{
computeMetrics(
  dataframeSequence,
  modality1,
  modality2,
  comparisonLabel = NULL,
  vars = Frame0:Frame100,
  colwise = FALSE,
  metric = "mae"
)
}
\arguments{
\item{dataframeSequence}{the Dataframe Sequence}

\item{modality1}{modality 1 of the column Comparison Factor}

\item{modality2}{modality 2 of the column Comparison Factor}

\item{comparisonLabel}{name of the comparison}

\item{vars}{variables to compare}

\item{colwise}{return metrics for each items of vars}

\item{metric}{metric name of the Metrics Package ( mae by default)}
}
\description{
compute metrics from the Metrics package
}
