% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeR2.R
\name{computeR2}
\alias{computeR2}
\title{compute of  the R-squared}
\usage{
computeR2(
  dataframeSequence,
  modality1,
  modality2,
  comparisonLabel = NULL,
  vars = Frame0:Frame100
)
}
\arguments{
\item{dataframeSequence}{the Dataframe Sequence}

\item{modality1}{modality 1 of the column Comparison Factor}

\item{modality2}{modality 2 of the column Comparison Factor}

\item{comparisonLabel}{name of the comparison}

\item{vars}{variables to compare}
}
\description{
compute of  the R-squared
}
