% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{ComputeSessionAverage}
\alias{ComputeSessionAverage}
\title{ComputeSessionAverage}
\usage{
ComputeSessionAverage(table)
}
\arguments{
\item{table}{\link{dataframe} all-table}
}
\value{
\link{dataframe}
}
\description{
compute average of all sessions
}
\section{Warning}{

}

\examples{
TODO
}
