% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructOverallEvents.R
\name{constructOverallEvents}
\alias{constructOverallEvents}
\title{merge the left and right eventCOntext}
\usage{
constructOverallEvents(df, groupBy, removeFirstCharacter = TRUE)
}
\arguments{
\item{removeFirstCharacter}{}
}
\description{
merge the left and right eventCOntext
}
