% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructTableFromXls.R
\name{constructTableFromXls}
\alias{constructTableFromXls}
\title{Construct a dataframe from a list of xls files}
\usage{
constructTableFromXls(fullXlsFiles, sheet)
}
\arguments{
\item{fullXlsFiles}{a list of xls files}

\item{sheet}{sheetname to extract from the xls files}
}
\description{
Construct a dataframe from a list of xls files
}
