% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineComparisonfactor.R
\name{defineComparisonfactor}
\alias{defineComparisonfactor}
\title{define the comparison factor}
\usage{
defineComparisonfactor(table, groupBy)
}
\arguments{
\item{groupBy}{}
}
\value{

}
\description{
define the comparison factor
}
