% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataScalarSequence.R
\name{DescritiveScalarSequence}
\alias{DescritiveScalarSequence}
\title{compute the descriptive statistics of scalar values}
\usage{
DescritiveScalarSequence(table, dependantVariables, bySubjectFlag = TRUE)
}
\arguments{
\item{bySubjectFlag}{}
}
\description{
compute the descriptive statistics of scalar values
}
