% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_consistencyLine.R
\name{geom_consistencyLine}
\alias{geom_consistencyLine}
\title{Title}
\usage{
geom_consistencyLine(
  table,
  color = "ComparisonFactor",
  linetype = NULL,
  lineWidth = 0.5
)
}
\arguments{
\item{table}{a dataframeSequences  with all cycles frame sequence}

\item{lineWidth}{}
}
\value{

}
\description{
Title
}
