% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_corridor_ribbon.R
\name{geom_corridor_ribbon}
\alias{geom_corridor_ribbon}
\title{construction of the  standard deviation ribbon}
\usage{
geom_corridor_ribbon(descrFrames)
}
\arguments{
\item{descrFrames}{a descriptive dataFrame Sequence}
}
\value{

}
\description{
construction of the  standard deviation ribbon
}
