% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_descriptiveLine.R
\name{geom_descriptiveLine}
\alias{geom_descriptiveLine}
\title{Title}
\usage{
geom_descriptiveLine(
  descrFrames,
  color = "ComparisonFactor",
  linetype = NULL,
  lineWidth = 0.5
)
}
\arguments{
\item{descrFrames}{a descriptive dataFrameSequence}

\item{lineWidth}{}
}
\description{
Title
}
