% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_gaitPhaseLines.R
\name{geom_gaitPhaseLines}
\alias{geom_gaitPhaseLines}
\title{draw gait phase as vertical lines}
\usage{
geom_gaitPhaseLines(descrDataFrame, color = NULL)
}
\arguments{
\item{descrDataFrame}{a descriptive dataFrame Sequence}

\item{color}{}
}
\description{
draw gait phase as vertical lines
}
