% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_normative_ribbon.R
\name{geom_normative_ribbon}
\alias{geom_normative_ribbon}
\title{Title}
\usage{
geom_normative_ribbon(normativeDataSet)
}
\arguments{
\item{normativeDataSet}{the normative data set.}
}
\value{

}
\description{
Title
}
