% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotlayer.R
\name{ggplotLayer}
\alias{ggplotLayer}
\title{pyrCGM2 ggplot layer}
\usage{
ggplotLayer(
  iTitle = "",
  xLabel = "Time Normalized",
  yLabel = "Deg",
  legendPosition = "top"
)
}
\arguments{
\item{legendPosition}{}
}
\value{

}
\description{
pyrCGM2 ggplot layer
}
