% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadNormativeDataSet.R
\name{loadNormativeDataSet}
\alias{loadNormativeDataSet}
\title{load a normative dataset}
\usage{
loadNormativeDataSet(fullXlsFile, sheet)
}
\arguments{
\item{sheet}{}
}
\description{
load a normative dataset
}
