% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrameSequence.R
\name{DescritiveFrameSequence}
\alias{DescritiveFrameSequence}
\title{return descriptive stats}
\usage{
DescritiveFrameSequence(table, bySubjectFlag = TRUE)
}
\arguments{
\item{bySubjectFlag}{}
}
\description{
return descriptive stats
}
